Move new integration test into existing test class.
